Revert "Kchro3/start up"
