+++ TypeaheadAI/Actors/SpecialCopyActor.swift
@@ -65,7 +65,7 @@ actor SpecialCopyActor: CanSimulateCopy {
                             copiedText: copiedText,
                             incognitoMode: incognitoMode,
                             stream: true,
-                            streamHandler: self.modalManager.defaultStreamHandler,
+                            streamHandler: self.modalManager.defaultHandler,
                             completion: { _ in }
                         )
                     }
