+++ TypeaheadAI/ClientManager.swift
@@ -51,8 +51,8 @@ class ClientManager {
     private let session: URLSession
 
     private let apiUrl = URL(string: "https://typeahead-ai.fly.dev/get_response")!
-//    private let apiUrlStreaming = URL(string: "https://typeahead-ai.fly.dev/get_response_stream")!
-    private let apiUrlStreaming = URL(string: "http://localhost:8080/get_response_stream")!
+    private let apiUrlStreaming = URL(string: "https://typeahead-ai.fly.dev/get_response_stream")!
+//    private let apiUrlStreaming = URL(string: "http://localhost:8080/get_response_stream")!
 
     private let logger = Logger(
         subsystem: "ai.typeahead.TypeaheadAI",
