+++ TypeaheadAI/TypeaheadAI.xcdatamodeld/TypeaheadAI.xcdatamodel/contents
@@ -17,14 +17,6 @@
         <attribute name="id" optional="YES" attributeType="UUID" usesScalarValueType="NO"/>
         <attribute name="summary" optional="YES" attributeType="String"/>
     </entity>
-    <entity name="MessageEntry" representedClassName="MessageEntry" syncable="YES" codeGenerationType="class">
-        <attribute name="attributed" optional="YES" attributeType="Binary"/>
-        <attribute name="createdAt" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
-        <attribute name="id" optional="YES" attributeType="UUID" usesScalarValueType="NO"/>
-        <attribute name="isCurrentUser" optional="YES" attributeType="Boolean" usesScalarValueType="YES"/>
-        <attribute name="responseError" optional="YES" attributeType="String"/>
-        <attribute name="text" optional="YES" attributeType="String"/>
-    </entity>
     <entity name="PromptEntry" representedClassName="PromptEntry" syncable="YES" codeGenerationType="class">
         <attribute name="createdAt" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
         <attribute name="id" optional="YES" attributeType="UUID" usesScalarValueType="NO"/>
