+++ TypeaheadAI/TypeaheadAIApp.swift
@@ -32,7 +32,6 @@ final class AppState: ObservableObject {
     @Published var llamaModelManager = LlamaModelManager()
     @Published var modalManager: ModalManager
     @Published var clientManager: ClientManager
-    @Published var messageManager: MessageManager
     private let historyManager: HistoryManager
     private let appContextManager: AppContextManager
     private let memoManager: MemoManager
@@ -56,11 +55,10 @@ final class AppState: ObservableObject {
 
         // Initialize managers
         self.memoManager = MemoManager(context: context)
-        self.historyManager = HistoryManager(context: context) // NOTE: Deprecate this
-        self.messageManager = MessageManager(context: context)
+        self.historyManager = HistoryManager(context: context)
         self.promptManager = PromptManager(context: context)
         self.clientManager = ClientManager()
-        self.modalManager = ModalManager(context: context)
+        self.modalManager = ModalManager()
         self.appContextManager = AppContextManager()
 
         // Initialize actors
@@ -85,7 +83,6 @@ final class AppState: ObservableObject {
         self.clientManager.promptManager = promptManager
         self.clientManager.appContextManager = appContextManager
         self.modalManager.clientManager = clientManager
-        self.modalManager.messageManager = messageManager
 
         checkAndRequestNotificationPermissions()
