+++ TypeaheadAI/Views/Modal/ModalView.swift
@@ -8,6 +8,76 @@
 import SwiftUI
 import Markdown
 
+struct MessageView: View {
+    let message: Message
+    var onButtonDown: (() -> Void)?
+
+    init(
+        message: Message,
+        onButtonDown: (() -> Void)? = nil
+    ) {
+        self.message = message
+        self.onButtonDown = onButtonDown
+    }
+
+    var body: some View {
+        if let error = message.responseError, !message.isCurrentUser {
+            HStack {
+                Text(error)
+                    .foregroundColor(.primary)
+                    .textSelection(.enabled)
+                    .padding(.vertical, 8)
+                    .padding(.horizontal, 15)
+                    .background(
+                        RoundedRectangle(cornerRadius: 15)
+                            .fill(Color.red.opacity(0.4))
+                    )
+
+                Button(action: {
+                    onButtonDown?()
+                }, label: {
+                    Image(systemName: "arrow.counterclockwise")
+                })
+                .buttonStyle(.plain)
+
+                Spacer()
+            }
+            .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .leading)
+        } else if let attributed = message.attributed {
+            attributedView(results: attributed.results)
+                .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: message.isCurrentUser ? .trailing : .leading)
+        } else if message.text.isEmpty && !message.isCurrentUser {
+            Divider()
+        } else {
+            Text(message.text)
+                .foregroundColor(message.isCurrentUser ? .white : .primary)
+                .textSelection(.enabled)
+                .padding(.vertical, 8)
+                .padding(.horizontal, 15)
+                .background(
+                    RoundedRectangle(cornerRadius: 15)
+                        .fill(message.isCurrentUser ? Color.blue.opacity(0.8) : Color.clear)
+                )
+                .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: message.isCurrentUser ? .trailing : .leading)
+        }
+    }
+
+    func attributedView(results: [ParserResult]) -> some View {
+        VStack(alignment: .leading, spacing: 0) {
+            ForEach(results) { parsed in
+                if parsed.isCodeBlock {
+                    CodeBlockView(parserResult: parsed)
+                        .padding(.bottom, 24)
+                        .textSelection(.enabled)
+                } else {
+                    Text(parsed.attributedString)
+                        .textSelection(.enabled)
+                }
+            }
+        }
+    }
+}
+
 struct ModalView: View {
     @Binding var showModal: Bool
     @State var incognito: Bool
@@ -76,34 +146,24 @@ struct ModalView_Previews: PreviewProvider {
     @State static var showModal = true
 
     static var previews: some View {
-        let container = NSPersistentContainer(name: "TypeaheadAI")
-        container.persistentStoreDescriptions.first?.url = URL(fileURLWithPath: "/dev/null")
-        container.loadPersistentStores { _, error in
-            if let error = error as NSError? {
-                fatalError("Unresolved error \(error), \(error.userInfo)")
-            }
-        }
-
-        let context = container.viewContext
-
-        let modalManager = ModalManager(context: context)
+        let modalManager = ModalManager()
         modalManager.setText("hello world")
 
-        let modalManagerWithMessages = ModalManager(context: context)
+        let modalManagerWithMessages = ModalManager()
         modalManagerWithMessages.messages = [
-            Message(id: UUID(), createdAt: Date(), text: "hello world", isCurrentUser: false),
-            Message(id: UUID(), createdAt: Date(), text: "hello bot", isCurrentUser: true)
+            Message(id: UUID(), text: "hello world", isCurrentUser: false),
+            Message(id: UUID(), text: "hello bot", isCurrentUser: true)
         ]
 
-        let modalManagerWithErrors = ModalManager(context: context)
+        let modalManagerWithErrors = ModalManager()
         modalManagerWithErrors.messages = [
-            Message(id: UUID(), createdAt: Date(), text: "", isCurrentUser: false, responseError: "Request took too long"),
-            Message(id: UUID(), createdAt: Date(), text: "hello bot", isCurrentUser: true)
+            Message(id: UUID(), text: "", isCurrentUser: false, responseError: "Request took too long"),
+            Message(id: UUID(), text: "hello bot", isCurrentUser: true)
         ]
 
-        let modalManagerWithCodeblock = ModalManager(context: context)
+        let modalManagerWithCodeblock = ModalManager()
         modalManagerWithCodeblock.messages = [
-            Message(id: UUID(), createdAt: Date(), text: markdownString, attributed: AttributedOutput(string: markdownString, results: [parserResult]), isCurrentUser: false)
+            Message(id: UUID(), text: markdownString, attributed: AttributedOutput(string: markdownString, results: [parserResult]), isCurrentUser: false)
         ]
 
         return Group {
